fix NumberFormatException
fix NumberFormatException for files >2,14 GB (=limit that can be represented by Int)

+++ app/src/main/java/com/nextcloud/talk/repositories/SharedItem.kt
@@ -5,7 +5,7 @@ import com.nextcloud.talk.models.database.UserEntity
 data class SharedItem(
     val id: String,
     val name: String,
-    val fileSize: Int,
+    val fileSize: Long,
     val date: Long,
     val path: String,
     val link: String,
